Move Model namespace to Models.
